# Conflicts:
#	Final/.gradle/6.5/executionHistory/executionHistory.bin
#	Final/.gradle/6.5/executionHistory/executionHistory.lock
#	Final/.gradle/6.5/fileHashes/fileHashes.bin
#	Final/.gradle/6.5/fileHashes/fileHashes.lock
#	Final/.gradle/6.5/fileHashes/resourceHashesCache.bin
#	Final/.gradle/6.5/javaCompile/classAnalysis.bin
#	Final/.gradle/6.5/javaCompile/jarAnalysis.bin
#	Final/.gradle/6.5/javaCompile/javaCompile.lock
#	Final/.gradle/6.5/javaCompile/taskHistory.bin
#	Final/.gradle/buildOutputCleanup/buildOutputCleanup.lock

+++ Final/app/src/main/java/ca/ulaval/ima/mp/model/DataClass.kt
@@ -92,7 +92,7 @@ data class Creator(
 data class Review(
         @SerializedName("id")  val id:Int,
         @SerializedName("creator") val creator:Creator,
-        @SerializedName("stars") val stars:Int,
+        @SerializedName("stars") val stars:Float,
         @SerializedName("image") val image:String,
         @SerializedName("comment") val comment:String,
         @SerializedName("date") val date:String
